qASDASD

+++ integrador/lib/pages/paghome.dart
@@ -14,19 +14,15 @@ class Paghome extends StatefulWidget {
 
 class _PaghomeState extends State<Paghome> {
 
-    final user = FirebaseAuth.instance.currentUser!;
+  final user = FirebaseAuth.instance.currentUser!;
   final CollectionReference cliente = FirebaseFirestore.instance.collection('cliente');
-  late String rg;
+  String rg = "";
 
   Future<void> pegaid() async {
     List<String> ids = [];
 
-    try {
-      QuerySnapshot querySnapshot = await FirebaseFirestore.instance.collection('cliente').get();
-      ids = querySnapshot.docs.map((doc) => doc.id).toList();
-    } catch (e) {
-      print('Erro ao obter IDs: $e');
-    }
+    QuerySnapshot querySnapshot = await FirebaseFirestore.instance.collection('cliente').get();
+    ids = querySnapshot.docs.map((doc) => doc.id).toList();
 
     for (var name in ids) {
       var documentSnapshot = await FirebaseFirestore.instance.collection('cliente').doc(name).get();
@@ -45,7 +41,9 @@ class _PaghomeState extends State<Paghome> {
 
   Future<void> deleta() async {
     await user.delete();
-    await cliente.doc(rg).delete();
+    if(rg!=""){
+      await cliente.doc(rg).delete();
+    }
     Navigator.pushNamed(context, '/paglogin');
   }
 
@@ -57,7 +55,7 @@ class _PaghomeState extends State<Paghome> {
           FutureBuilder(
             future: initializeRg(),
             builder: (context, snapshot) { 
-              if (rg!=null){
+              if (rg!=""){
                 return Peganome(id: rg);
               }
               else{
